% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_layer.R
\name{extract_layer}
\alias{extract_layer}
\title{Extract Layer}
\usage{
extract_layer(res, layer_no = 0, folder_path = NULL)
}
\arguments{
\item{res}{A rest_response object returned by \code{\link{get_layers}}.}

\item{layer_no}{A single integer, 0 to n - 1, describing the layer to extract.}

\item{folder_path}{An optional path to a directory. If it exists, any attachments will be downloaded there.}
}
\value{
A data frame. If there were attachments, the URLs for each record will be aggregated in to a list and then added to the data frame as an ``url'' field. If the \code{folder_path} was specified, then instead, the paths of the attachments will be added as a ``path'' field.
}
\description{
Extracts a data frame from the layers returned by \code{\link{get_layers}}. Optionally, downloads the layer's attachments.
}
