% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_layer.R
\name{get_layer_info}
\alias{get_layer_info}
\alias{get_layer}
\title{Get Layer}
\usage{
get_layer_info(rest, layer_no)

get_layer(rest, layer_no)
}
\arguments{
\item{rest}{A \code{\link{rest_service}} object.}

\item{layer_no}{Scalar, the number of the layer, 0 to n - 1.}
}
\value{
\code{get_layer_info} returns a \code{rest_response} object. \code{get_layer}, the attribute table as a data frame.
}
\description{
Get a Layer from a REST Service
}
\section{Functions}{
\itemize{
\item \code{get_layer_info}: Get information about a layer.
}}

