% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_layers.R
\name{get_layers}
\alias{get_layers}
\title{Get Layers}
\usage{
get_layers(rest, layers = all_layers(rest))
}
\arguments{
\item{rest}{A \code{\link{rest_service}} object.}

\item{layers}{A character vector, the numbers of the layers, 0 to n - 1. Defaults to all of them.}
}
\value{
A \code{rest_response} object with the content normalized.
}
\description{
Get one or more layers from a REST Service
}
