% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_token.R
\name{get_token}
\alias{get_token}
\title{Get Token}
\usage{
get_token(usr, pwd)
}
\arguments{
\item{usr}{Character, the user name.}

\item{pwd}{Character, the password.}
}
\value{
Character, the value of the token.
}
\description{
Get an authentication token, reusing the last one if it hasn't expired.
}
\examples{
\dontrun{get_token("user", "password")}

}
