% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rest_response.R
\name{print.rest_response}
\alias{print.rest_response}
\title{Print a Rest Response}
\usage{
\method{print}{rest_response}(x, ...)
}
\arguments{
\item{...}{Ignored.}

\item{The}{object.}
}
\description{
S3 method to print a \code{rest_response} object.
}
