% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcRest.R
\name{print.rest_service}
\alias{print.rest_service}
\title{Print a Rest Service}
\usage{
\method{print}{rest_service}(x, ...)
}
\arguments{
\item{...}{Ignored.}

\item{The}{object.}
}
\description{
S3 method to print a \code{rest_service} object.
}
