% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rest_response.R
\name{rest_GET}
\alias{rest_GET}
\alias{rest_POST}
\title{REST Get}
\usage{
rest_GET(url, usr, pwd, query = list())

rest_POST(url, usr, pwd, query = list(), body = list())
}
\arguments{
\item{url}{The url of the resource being requested.}

\item{usr}{Character, the user name.}

\item{pwd}{Character, the password.}

\item{query}{A list of key value pairs to be sent with URL of the request.}

\item{body}{A list of key value pairs to be sent with HTTP request headers.}
}
\value{
A list of class \code{rest_response} with the members:

\describe{
\item{url}{The URL of the request.}
\item{query}{The key value pairs passed in the URL.}
\item{body}{The key value pairs passed in the request headers.}
\item{content}{The parsed content returned by the REST service.}
}
}
\description{
Send an HTTP request to a REST service.
}
\section{Functions}{
\itemize{
\item \code{rest_POST}: Send an HTTP put request to a REST service.
}}

\section{Note}{

These functions negotiate the authentication token, caching it in memory and fetching a new one when when necessary.
}

