% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcRest.R
\name{rest_service}
\alias{rest_service}
\title{Create a Rest Service object.}
\usage{
rest_service(url, usr, pwd)
}
\arguments{
\item{url}{The URL of the service.}

\item{usr}{Character, the user name.}

\item{pwd}{Character, the password.}
}
\value{
A list of class \code{rest_service} with the members:

\describe{
\item{url}{The URL of the service. This may be a truncated version of what was used to construct the object.}
\item{usr}{The user name for the REST service.}
\item{pwd}{The password for the REST service.}
\item{content}{The parsed content returned by the REST service.}
}
}
\description{
Create a Rest Service object.
}
