% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_layers.R
\name{get_layer_info}
\alias{get_layer_info}
\alias{get_layers}
\title{Get Layers}
\usage{
get_layer_info(rest, layer_no)

get_layers(rest, layers = all_layers(rest))
}
\arguments{
\item{rest}{A \code{\link{rest_service}} object.}

\item{layer_no}{An integer, the number of the layer to get information about.}

\item{layers}{A character vector, the numbers of the layers, 0 to n - 1. Defaults to all of them.}

\item{rest}{A \code{\link{rest_service}} object.}
}
\value{
A \code{rest_response} object with the content normalized.
}
\description{
Get one or more layers from a REST Service
}
\section{Functions}{
\itemize{
\item \code{get_layer_info}: Get information about a layer.
}}

