% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rest_response.R
\name{rest_GET}
\alias{rest_GET}
\alias{rest_POST}
\alias{rest_download}
\title{REST Get}
\usage{
rest_GET(rest, path, params = list())

rest_POST(rest, path, params = list())

rest_download(rest, path = NULL, url = NULL, destfile,
  params = list())
}
\arguments{
\item{rest}{A rest_service object returned by \code{\link{rest_service}}.}

\item{path}{The path of the resource being requested, relative to the URL of the rest service.}

\item{params}{A list of key value pairs to be sent with the request.}

\item{url}{The full URL of the file being requested, absent of the query part.}
}
\value{
A list of class \code{rest_response} with the members:

\describe{
\item{rest}{The \code{rest_service} object associated with the request.}
\item{path}{The path of the request.}
\item{params}{The key value pairs passed in the request.}
\item{content}{The parsed content returned by the REST service. Or in the case of \code{rest_download}, the response form \code{file.info} but with the file path appended.}
}
}
\description{
Send an HTTP request to a REST service.
}
\section{Functions}{
\itemize{
\item \code{rest_POST}: Send an HTTP put request to a REST service.

\item \code{rest_download}: Send an HTTP put request to a REST service.
}}

\section{Note}{

These functions negotiate the authentication token, caching it in memory and fetching a new one when when necessary.
}

